Rename WebsocketAdapter to RemoteConsole::ServerAdapter with children
