+++ lib/remote_console/server_adapter/web_mks_legacy.rb
@@ -1,0 +1,22 @@
+module RemoteConsole
+  module ServerAdapter
+    class WebMKSLegacy < Websocket
+      def fetch(length)
+        @driver.on(:message) { |msg| yield(msg.data) } if @driver.listeners(:message).length.zero?
+
+        data = @sock.read_nonblock(length)
+        @driver.parse(data)
+      end
+
+      def issue(data)
+        @driver.frame(data)
+      end
+
+      private
+
+      def protocol
+        'uint8utf8'.freeze
+      end
+    end
+  end
+end
